+++ tests/test_transform.py
@@ -1,18 +1,16 @@
 import csv
 from datetime import datetime
+import dateutil
 import glob
 import io
 import json
 import os
 import unittest
-from unittest.mock import patch
 import zipfile
 
-import dateutil
-
 from transform import app
-from transform.transformers.ImageTransformer import ImageTransformer
 from transform.views.image_filters import format_date
+from unittest.mock import patch
 
 
 def get_file_as_string(filename):
@@ -121,7 +119,7 @@ class TestTransformService(unittest.TestCase):
 
             self.assertEqual(actual_response, expected_response)
 
-    @patch('transform.transformers.ImageTransformer.get_image_sequence_numbers', return_value=[1, 2])
+    @patch('transform.transformers.ImageTransformer._get_image_sequence_list', return_value=[1, 2])
     def test_transforms_csv(self, mock_sequence_no):
         test_scenarios = get_test_scenarios('csv')
 
@@ -188,12 +186,3 @@ class TestTransformService(unittest.TestCase):
         r = self.app.post(self.transform_pck_endpoint, data=payload)
 
         self.assertEqual(r.status_code, 400)
-
-    def test_cleanup(self):
-        for dirpath, dirnames, files in os.walk('./tmp'):
-            if dirnames:
-                ImageTransformer.cleanup(self, './tmp')
-            if not dirnames:
-                self.assertEqual(dirpath, './tmp')
-                self.assertEqual(dirnames, [])
-                self.assertEqual(files, [])