+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### Unreleased
+  - Move transformers to be memory based instead of file based
+
  - Initial files for move to cloudfoundry
 ### 2.12.0 2017-11-09
   - Fix incorrect index date time